requirements.txt
aniso8601==6.0.0
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.3
coveralls==1.8.1
dj-database-url==0.5.0
django-graphql-jwt==0.2.1
django-heroku==0.3.1
django-nose==1.4.6
django==2.2.4
docopt==0.6.2
factory-boy==2.12.0
Faker==2.0.0
future==0.17.1
graphene-django==2.3.2
graphene==2.1.6
graphql-core==2.2
graphql-relay==0.4.5
gunicorn==19.9.0
idna==2.8
nose==1.3.7
pep8==1.7.1
promise==2.2.1
psycopg2-binary==2.8.3
psycopg2==2.7.7
pycodestyle==2.5.0
PyJWT==1.7.1
python-dateutil==2.8.0
python-decouple==3.1
python-http-client==3.1.0
pytz==2019.1
requests==2.22.0
rx==1.6.1
sendgrid==6.0.5
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
urllib3==1.25.3
whitenoise==4.1.2
